automod: extract countstore to package.
Also separate the mem implementation and the interface into their own
files.

No other changes in this commit but moves, to keep history as legible
as possible.

This is just a feature grouping package, to reduce how much is in the
main automod package.  Implementations are not separted into further
subpackages because there's no visible need to do that as yet (e.g.
we're not particular concerned with minimizing dep graphs, etc).

Some other parts of automod might deserve similar extraction (the
cachestore is a likely example) but that can be pursued separately.
Yet others (namely, the redis identity directory) might deserve
extraction, but that one might already be general well beyond
automod, perhaps deserving to live out in the atproto packages,
so that will definitely deserve its own full review.
