+++ automod/countstore/countstore_mem.go
@@ -1,4 +1,4 @@
-package automod
+package countstore
 
 import (
 	"context"
@@ -6,20 +6,6 @@ import (
 	"log/slog"
 	"time"
 )
-
-const (
-	PeriodTotal = "total"
-	PeriodDay   = "day"
-	PeriodHour  = "hour"
-)
-
-type CountStore interface {
-	GetCount(ctx context.Context, name, val, period string) (int, error)
-	Increment(ctx context.Context, name, val string) error
-	// TODO: batch increment method
-	GetCountDistinct(ctx context.Context, name, bucket, period string) (int, error)
-	IncrementDistinct(ctx context.Context, name, bucket, val string) error
-}
 
 // TODO: this implementation isn't race-safe (yet)!
 type MemCountStore struct {
