+++ automod/countstore/countstore_test.go
@@ -1,4 +1,4 @@
-package automod
+package countstore
 
 import (
 	"context"
