+++ automod/engine.go
@@ -9,6 +9,7 @@ import (
 	comatproto "github.com/bluesky-social/indigo/api/atproto"
 	"github.com/bluesky-social/indigo/atproto/identity"
 	"github.com/bluesky-social/indigo/atproto/syntax"
+	"github.com/bluesky-social/indigo/automod/countstore"
 	"github.com/bluesky-social/indigo/xrpc"
 )
 
@@ -19,7 +20,7 @@ type Engine struct {
 	Logger      *slog.Logger
 	Directory   identity.Directory
 	Rules       RuleSet
-	Counters    CountStore
+	Counters    countstore.CountStore
 	Sets        SetStore
 	Cache       CacheStore
 	Flags       FlagStore
