+++ automod/engine_test.go
@@ -9,6 +9,7 @@ import (
 	appbsky "github.com/bluesky-social/indigo/api/bsky"
 	"github.com/bluesky-social/indigo/atproto/identity"
 	"github.com/bluesky-social/indigo/atproto/syntax"
+	"github.com/bluesky-social/indigo/automod/countstore"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -54,7 +55,7 @@ func engineFixture() Engine {
 	engine := Engine{
 		Logger:    slog.Default(),
 		Directory: &dir,
-		Counters:  NewMemCountStore(),
+		Counters:  countstore.NewMemCountStore(),
 		Sets:      sets,
 		Flags:     flags,
 		Cache:     cache,
